# perf: use fmt.Fprintf to avoid unnecessary string+args + WriteString